// src/app/core/models/prefix/prefix.model.ts
export interface Prefix {
  branch: boolean;
  codeName: string;
  createdBy: string;
  createdOn: string;
  dept: boolean;
  financialYear: boolean;
  firstCustomValue: string;
  fyChangedBy: string;
  fyChangedOn: string;
  id: number;
  lastCustomValue: string;
  modifiedBy: string;
  modifiedOn: string;
  transactionType: boolean;
}